Добавляет четвертое домашнее задание
